Extract all knex calls behind a Driver interface.
Also pulled all the dataloader logic out behind the Driver interface
as well, b/c in theory an in-memory driver wouldn't require batching.

...although would that cause nuanced runtime variables between the
drivers?

Basically all drivers would want to share that logic too, so seems like
that's not the right arrangement.
