+++ packages/integration-tests/src/entities/AuthorCodegen.ts
@@ -1,51 +1,51 @@
 import {
-  BaseEntity,
-  BooleanFilter,
-  BooleanGraphQLFilter,
-  Changes,
-  Collection,
-  ConfigApi,
-  EntityFilter,
-  EntityGraphQLFilter,
-  EntityManager,
-  FilterOf,
   Flavor,
-  getEm,
-  GraphQLFilterOf,
-  hasMany,
-  hasOne,
-  hasOneToOne,
-  Lens,
-  Loaded,
-  LoadHint,
-  loadLens,
-  newChangesProxy,
-  newRequiredRule,
-  OptsOf,
-  OrderBy,
-  PartialOrNull,
-  Reference,
-  setField,
-  setOpts,
   ValueFilter,
   ValueGraphQLFilter,
+  OrderBy,
+  ConfigApi,
+  BaseEntity,
+  EntityManager,
+  setOpts,
+  PartialOrNull,
+  OptsOf,
+  Changes,
+  newChangesProxy,
+  Lens,
+  loadLens,
+  LoadHint,
+  Loaded,
+  getEm,
+  BooleanFilter,
+  EntityFilter,
+  FilterOf,
+  BooleanGraphQLFilter,
+  EntityGraphQLFilter,
+  GraphQLFilterOf,
+  newRequiredRule,
+  Collection,
+  hasMany,
+  Reference,
+  hasOne,
+  hasOneToOne,
+  setField,
 } from "joist-orm";
-import { Context } from "src/context";
 import {
   Author,
-  authorMeta,
-  Book,
-  BookId,
-  bookMeta,
-  Image,
-  ImageId,
-  imageMeta,
   newAuthor,
+  authorMeta,
   Publisher,
+  Image,
+  Book,
   PublisherId,
-  publisherMeta,
+  ImageId,
+  BookId,
   PublisherOrder,
+  bookMeta,
+  publisherMeta,
+  imageMeta,
 } from "./entities";
+import { Context } from "src/context";
 
 export type AuthorId = Flavor<string, "Author">;
