+++ packages/integration-tests/src/entities/BookAdvanceCodegen.ts
@@ -1,46 +1,46 @@
 import {
-  BaseEntity,
-  Changes,
-  ConfigApi,
-  EntityFilter,
-  EntityGraphQLFilter,
-  EntityManager,
-  EnumGraphQLFilter,
-  FilterOf,
   Flavor,
-  getEm,
-  GraphQLFilterOf,
-  hasOne,
-  Lens,
-  Loaded,
-  LoadHint,
-  loadLens,
-  newChangesProxy,
-  newRequiredRule,
-  OptsOf,
-  OrderBy,
-  PartialOrNull,
-  Reference,
-  setField,
-  setOpts,
   ValueFilter,
   ValueGraphQLFilter,
+  OrderBy,
+  ConfigApi,
+  BaseEntity,
+  EntityManager,
+  setOpts,
+  PartialOrNull,
+  OptsOf,
+  Changes,
+  newChangesProxy,
+  Lens,
+  loadLens,
+  LoadHint,
+  Loaded,
+  getEm,
+  EntityFilter,
+  FilterOf,
+  EnumGraphQLFilter,
+  EntityGraphQLFilter,
+  GraphQLFilterOf,
+  newRequiredRule,
+  setField,
+  Reference,
+  hasOne,
 } from "joist-orm";
-import { Context } from "src/context";
 import {
+  BookAdvance,
+  newBookAdvance,
+  bookAdvanceMeta,
   AdvanceStatus,
   Book,
-  BookAdvance,
-  bookAdvanceMeta,
-  BookId,
-  bookMeta,
-  BookOrder,
-  newBookAdvance,
   Publisher,
+  BookId,
   PublisherId,
-  publisherMeta,
+  BookOrder,
   PublisherOrder,
+  bookMeta,
+  publisherMeta,
 } from "./entities";
+import { Context } from "src/context";
 
 export type BookAdvanceId = Flavor<string, "BookAdvance">;
