+++ packages/integration-tests/src/entities/BookCodegen.ts
@@ -1,56 +1,56 @@
 import {
-  BaseEntity,
-  Changes,
-  Collection,
-  ConfigApi,
-  EntityFilter,
-  EntityGraphQLFilter,
-  EntityManager,
-  FilterOf,
   Flavor,
-  getEm,
-  GraphQLFilterOf,
-  hasMany,
-  hasManyToMany,
-  hasOne,
-  hasOneToOne,
-  Lens,
-  Loaded,
-  LoadHint,
-  loadLens,
-  newChangesProxy,
-  newRequiredRule,
-  OptsOf,
-  OrderBy,
-  PartialOrNull,
-  Reference,
-  setField,
-  setOpts,
   ValueFilter,
   ValueGraphQLFilter,
+  OrderBy,
+  ConfigApi,
+  BaseEntity,
+  EntityManager,
+  setOpts,
+  PartialOrNull,
+  OptsOf,
+  Changes,
+  newChangesProxy,
+  Lens,
+  loadLens,
+  LoadHint,
+  Loaded,
+  getEm,
+  EntityFilter,
+  FilterOf,
+  EntityGraphQLFilter,
+  GraphQLFilterOf,
+  newRequiredRule,
+  Collection,
+  hasMany,
+  Reference,
+  hasOne,
+  hasOneToOne,
+  hasManyToMany,
+  setField,
 } from "joist-orm";
-import { Context } from "src/context";
 import {
-  Author,
-  AuthorId,
-  authorMeta,
-  AuthorOrder,
   Book,
-  BookAdvance,
-  BookAdvanceId,
-  bookAdvanceMeta,
-  bookMeta,
-  BookReview,
-  BookReviewId,
-  bookReviewMeta,
-  Image,
-  ImageId,
-  imageMeta,
   newBook,
+  bookMeta,
+  Author,
+  Image,
+  BookAdvance,
+  BookReview,
   Tag,
+  AuthorId,
+  ImageId,
+  BookAdvanceId,
+  BookReviewId,
   TagId,
+  AuthorOrder,
+  bookAdvanceMeta,
+  bookReviewMeta,
+  authorMeta,
+  imageMeta,
   tagMeta,
 } from "./entities";
+import { Context } from "src/context";
 
 export type BookId = Flavor<string, "Book">;
