+++ packages/integration-tests/src/entities/BookReviewCodegen.ts
@@ -1,34 +1,34 @@
 import {
-  BaseEntity,
-  BooleanFilter,
-  BooleanGraphQLFilter,
-  Changes,
-  ConfigApi,
-  EntityFilter,
-  EntityGraphQLFilter,
-  EntityManager,
-  FilterOf,
   Flavor,
-  getEm,
-  GraphQLFilterOf,
-  hasOne,
-  Lens,
-  Loaded,
-  LoadHint,
-  loadLens,
-  newChangesProxy,
-  newRequiredRule,
-  OptsOf,
-  OrderBy,
-  PartialOrNull,
-  Reference,
-  setField,
-  setOpts,
   ValueFilter,
   ValueGraphQLFilter,
+  OrderBy,
+  ConfigApi,
+  BaseEntity,
+  EntityManager,
+  setOpts,
+  PartialOrNull,
+  OptsOf,
+  Changes,
+  newChangesProxy,
+  Lens,
+  loadLens,
+  LoadHint,
+  Loaded,
+  getEm,
+  BooleanFilter,
+  EntityFilter,
+  FilterOf,
+  BooleanGraphQLFilter,
+  EntityGraphQLFilter,
+  GraphQLFilterOf,
+  newRequiredRule,
+  Reference,
+  hasOne,
+  setField,
 } from "joist-orm";
+import { BookReview, newBookReview, bookReviewMeta, Book, BookId, BookOrder, bookMeta } from "./entities";
 import { Context } from "src/context";
-import { Book, BookId, bookMeta, BookOrder, BookReview, bookReviewMeta, newBookReview } from "./entities";
 
 export type BookReviewId = Flavor<string, "BookReview">;
