+++ packages/integration-tests/src/entities/ImageCodegen.ts
@@ -1,50 +1,50 @@
 import {
-  BaseEntity,
-  Changes,
-  ConfigApi,
-  EntityFilter,
-  EntityGraphQLFilter,
-  EntityManager,
-  EnumGraphQLFilter,
-  FilterOf,
   Flavor,
-  getEm,
-  GraphQLFilterOf,
-  hasOne,
-  Lens,
-  Loaded,
-  LoadHint,
-  loadLens,
-  newChangesProxy,
-  newRequiredRule,
-  OptsOf,
-  OrderBy,
-  PartialOrNull,
-  Reference,
-  setField,
-  setOpts,
   ValueFilter,
   ValueGraphQLFilter,
+  OrderBy,
+  ConfigApi,
+  BaseEntity,
+  EntityManager,
+  setOpts,
+  PartialOrNull,
+  OptsOf,
+  Changes,
+  newChangesProxy,
+  Lens,
+  loadLens,
+  LoadHint,
+  Loaded,
+  getEm,
+  EntityFilter,
+  FilterOf,
+  EnumGraphQLFilter,
+  EntityGraphQLFilter,
+  GraphQLFilterOf,
+  newRequiredRule,
+  setField,
+  Reference,
+  hasOne,
 } from "joist-orm";
-import { Context } from "src/context";
 import {
-  Author,
-  AuthorId,
-  authorMeta,
-  AuthorOrder,
-  Book,
-  BookId,
-  bookMeta,
-  BookOrder,
   Image,
+  newImage,
   imageMeta,
   ImageType,
-  newImage,
+  Author,
+  Book,
   Publisher,
+  AuthorId,
+  BookId,
   PublisherId,
-  publisherMeta,
+  AuthorOrder,
+  BookOrder,
   PublisherOrder,
+  authorMeta,
+  bookMeta,
+  publisherMeta,
 } from "./entities";
+import { Context } from "src/context";
 
 export type ImageId = Flavor<string, "Image">;
