+++ packages/integration-tests/src/entities/PublisherCodegen.ts
@@ -1,43 +1,43 @@
 import {
-  BaseEntity,
-  Changes,
-  Collection,
-  ConfigApi,
-  EntityManager,
-  EnumGraphQLFilter,
   Flavor,
-  getEm,
-  hasMany,
-  Lens,
-  Loaded,
-  LoadHint,
-  loadLens,
-  newChangesProxy,
-  newRequiredRule,
-  OptsOf,
-  OrderBy,
-  PartialOrNull,
-  setField,
-  setOpts,
   ValueFilter,
   ValueGraphQLFilter,
+  OrderBy,
+  ConfigApi,
+  BaseEntity,
+  EntityManager,
+  setOpts,
+  PartialOrNull,
+  OptsOf,
+  Changes,
+  newChangesProxy,
+  Lens,
+  loadLens,
+  LoadHint,
+  Loaded,
+  getEm,
+  EnumGraphQLFilter,
+  newRequiredRule,
+  setField,
+  Collection,
+  hasMany,
 } from "joist-orm";
-import { Context } from "src/context";
 import {
-  Author,
-  AuthorId,
-  authorMeta,
-  BookAdvance,
-  BookAdvanceId,
-  bookAdvanceMeta,
-  Image,
-  ImageId,
-  imageMeta,
-  newPublisher,
   Publisher,
+  newPublisher,
   publisherMeta,
   PublisherSize,
+  Author,
+  BookAdvance,
+  Image,
+  AuthorId,
+  BookAdvanceId,
+  ImageId,
+  authorMeta,
+  bookAdvanceMeta,
+  imageMeta,
 } from "./entities";
+import { Context } from "src/context";
 
 export type PublisherId = Flavor<string, "Publisher">;
