+++ packages/integration-tests/src/entities/metadata.ts
@@ -1,39 +1,39 @@
 import {
-  configureMetadata,
-  DecimalToNumberSerde,
   EntityManager as EntityManager1,
+  configureMetadata,
   EntityMetadata,
-  EnumFieldSerde,
-  ForeignKeySerde,
   PrimaryKeySerde,
   SimpleSerde,
+  ForeignKeySerde,
+  EnumFieldSerde,
+  DecimalToNumberSerde,
 } from "joist-orm";
 import { Context } from "src/context";
 import {
-  AdvanceStatuses,
   Author,
   authorConfig,
+  newAuthor,
   Book,
+  bookConfig,
+  newBook,
   BookAdvance,
   bookAdvanceConfig,
-  bookConfig,
+  newBookAdvance,
   BookReview,
   bookReviewConfig,
+  newBookReview,
   Image,
   imageConfig,
-  ImageTypes,
-  newAuthor,
-  newBook,
-  newBookAdvance,
-  newBookReview,
   newImage,
-  newPublisher,
-  newTag,
   Publisher,
   publisherConfig,
-  PublisherSizes,
+  newPublisher,
   Tag,
   tagConfig,
+  newTag,
+  AdvanceStatuses,
+  ImageTypes,
+  PublisherSizes,
 } from "./entities";
 
 export class EntityManager extends EntityManager1<Context> {}
