+++ packages/integration-tests/src/resolvers/enumResolvers.ts
@@ -1,5 +1,5 @@
-import { AdvanceStatuses, ImageTypes, PublisherSizes } from "src/entities";
 import { Resolvers } from "src/generated/graphql-types";
+import { AdvanceStatuses, ImageTypes, PublisherSizes } from "src/entities";
 
 type EnumDetails = "AdvanceStatusDetail" | "ImageTypeDetail" | "PublisherSizeDetail";
