+++ packages/orm/src/collections/OneToManyCollection.ts
@@ -1,8 +1,5 @@
-import DataLoader from "dataloader";
 import {
-  assertIdsAreTagged,
   Collection,
-  deTagIds,
   ensureNotDeleted,
   Entity,
   EntityMetadata,
@@ -11,7 +8,7 @@ import {
   IdOf,
   maybeResolveReferenceToId,
 } from "../index";
-import { getOrSet, groupBy, remove } from "../utils";
+import { remove } from "../utils";
 import { AbstractRelationImpl } from "./AbstractRelationImpl";
 import { ManyToOneReference } from "./ManyToOneReference";
 
@@ -45,7 +42,8 @@ export class OneToManyCollection<T extends Entity, U extends Entity>
       if (this.entity.id === undefined) {
         this.loaded = [];
       } else {
-        this.loaded = await loaderForCollection(this).load(this.entity.id);
+        const em = getEm(this.entity);
+        this.loaded = await em.driver.loadOneToMany(em, this);
       }
       this.maybeAppendAddedBeforeLoaded();
     }
@@ -159,9 +157,8 @@ export class OneToManyCollection<T extends Entity, U extends Entity>
   async refreshIfLoaded(): Promise<void> {
     // TODO We should remember what load hints have been applied to this collection and re-apply them.
     if (this.loaded !== undefined && this.entity.id !== undefined) {
-      const loader = loaderForCollection(this);
-      loader.clear(this.entity.id);
-      this.loaded = await loader.load(this.entity.id);
+      const em = getEm(this.entity);
+      this.loaded = await em.driver.loadOneToMany(em, this);
     }
   }
 
@@ -212,44 +209,4 @@ export class OneToManyCollection<T extends Entity, U extends Entity>
   private getOtherRelation(other: U): ManyToOneReference<U, T, any> {
     return (other as U)[this.otherFieldName] as any;
   }
-}
-
-function loaderForCollection<T extends Entity, U extends Entity>(
-  collection: OneToManyCollection<T, U>,
-): DataLoader<string, U[]> {
-  const em = getEm(collection.entity);
-  // The metadata for the entity that contains the collection
-  const meta = getMetadata(collection.entity);
-  const loaderName = `${meta.tableName}.${collection.fieldName}`;
-  return getOrSet(em.__data.loaders, loaderName, () => {
-    return new DataLoader<string, U[]>(async (_keys) => {
-      const { otherMeta } = collection;
-
-      assertIdsAreTagged(_keys);
-      const keys = deTagIds(meta, _keys);
-
-      const rows = await em.knex
-        .select("*")
-        .from(otherMeta.tableName)
-        .whereIn(collection.otherColumnName, keys)
-        .orderBy("id");
-
-      const entities = rows.map((row) => em.hydrate(otherMeta.cstr, row, { overwriteExisting: false }));
-      // .filter((e) => !e.isDeletedEntity);
-
-      const rowsById = groupBy(entities, (entity) => {
-        // TODO If this came from the UoW, it may not be an id? I.e. pre-insert.
-        const ownerId = maybeResolveReferenceToId(entity.__orm.data[collection.otherFieldName]);
-        // We almost always expect ownerId to be found, b/c normally we just hydrated this entity
-        // directly from a SQL row with owner_id=X, however we might be loading this collection
-        // (i.e. find all children where owner_id=X) when the SQL thinks a child is still pointing
-        // at the parent (i.e. owner_id=X in the db), but our already-loaded child has had its
-        // `child.owner` field either changed to some other owner, or set to undefined. In either,
-        // that child should no longer be parent of this owner's collection, so just return a
-        // dummy value.
-        return ownerId ?? "dummyNoLongerOwned";
-      });
-      return _keys.map((k) => rowsById.get(k) || []);
-    });
-  });
 }
