+++ packages/orm/src/drivers/EntityPersister.ts
@@ -1,15 +1,30 @@
+import DataLoader from "dataloader";
 import Knex from "knex";
 import { JoinRow, ManyToManyCollection } from "../collections/ManyToManyCollection";
 import {
+  afterTransaction,
+  beforeTransaction,
   deTagIds,
   Entity,
+  EntityConstructor,
+  EntityManager,
   EntityMetadata,
+  FilterOf,
   getMetadata,
   keyToNumber,
   keyToString,
   maybeResolveReferenceToId,
+  OneToManyCollection,
+  OneToOneReference,
+  OrderOf,
 } from "../index";
 import { partition } from "../utils";
+import { Driver } from "./driver";
+import { loaderForFind } from "./findDataLoader";
+import { loaderForJoinTable } from "./joinTableDataLoader";
+import { loaderForEntity } from "./loadDataLoader";
+import { loaderForCollection } from "./oneToManyDataLoader";
+import { loaderForOneToOne } from "./oneToOneDataLoader";
 
 /** The operations for a given entity type, so they can be executed in bulk. */
 export interface Todo {
@@ -20,21 +35,152 @@ export interface Todo {
   validates: Entity[];
 }
 
-export async function flushEntities(knex: Knex, todos: Record<string, Todo>): Promise<void> {
-  const updatedAt = new Date();
-  await assignNewIds(knex, todos);
-  for await (const todo of Object.values(todos)) {
-    if (todo) {
-      const meta = todo.metadata;
-      if (todo.inserts.length > 0) {
-        await batchInsert(knex, meta, todo.inserts);
+export type LoaderCache = Record<string, DataLoader<any, any>>;
+
+export class EntityPersister implements Driver {
+  private loadLoaders: LoaderCache = {};
+  private findLoaders: LoaderCache = {};
+
+  constructor(private knex: Knex) {}
+
+  resetDataLoaderCache() {
+    this.loadLoaders = {};
+    this.findLoaders = {};
+  }
+
+  load<T extends Entity>(em: EntityManager, meta: EntityMetadata<T>, id: string): Promise<T | undefined> {
+    return loaderForEntity(em, this.knex, this.loadLoaders, meta).load(id);
+  }
+
+  loadManyToMany<T extends Entity, U extends Entity>(
+    em: EntityManager,
+    collection: ManyToManyCollection<T, U>,
+  ): Promise<U[]> {
+    const { columnName, entity } = collection;
+    const key = `${columnName}=${entity.id}`;
+    return loaderForJoinTable(this.knex, this.loadLoaders, collection).load(key);
+  }
+
+  loadOneToMany<T extends Entity, U extends Entity>(
+    em: EntityManager,
+    collection: OneToManyCollection<T, U>,
+  ): Promise<U[]> {
+    return loaderForCollection(this.knex, this.loadLoaders, collection).load(collection.entity.id!);
+  }
+
+  async loadOneToOne<T extends Entity, U extends Entity>(
+    em: EntityManager,
+    reference: OneToOneReference<T, U>,
+  ): Promise<U | undefined> {
+    return (await loaderForOneToOne(this.knex, this.loadLoaders, reference).load(reference.entity.idOrFail))[0];
+  }
+
+  find<T extends Entity>(
+    em: EntityManager,
+    type: EntityConstructor<T>,
+    where: FilterOf<T>,
+    options?: { orderBy?: OrderOf<T>; limit?: number; offset?: number },
+  ): Promise<unknown[]> {
+    return loaderForFind(this.knex, this.findLoaders, type).load({ where, ...options });
+  }
+
+  async transaction<T>(
+    em: EntityManager,
+    fn: (txn: Knex.Transaction) => Promise<T>,
+    isolationLevel?: "serializable",
+  ): Promise<T> {
+    const alreadyInTxn = "commit" in this.knex;
+    if (alreadyInTxn) {
+      return fn(this.knex as Knex.Transaction);
+    }
+    const originalKnex = this.knex;
+    const txn = await this.knex.transaction();
+    this.knex = txn;
+    try {
+      if (isolationLevel) {
+        await txn.raw("set transaction isolation level serializable;");
       }
-      if (todo.updates.length > 0) {
-        todo.updates.forEach((e) => (e.__orm.data["updatedAt"] = updatedAt));
-        await batchUpdate(knex, meta, todo.updates);
+      await beforeTransaction(em, txn);
+      const result = await fn(txn);
+      await txn.commit();
+      await afterTransaction(em, txn);
+      return result;
+    } finally {
+      if (!txn.isCompleted()) {
+        txn.rollback().catch((e) => {
+          console.error(e, "Error rolling back");
+        });
       }
-      if (todo.deletes.length > 0) {
-        await batchDelete(knex, meta, todo.deletes);
+      this.knex = originalKnex;
+    }
+  }
+
+  async flushEntities(todos: Record<string, Todo>): Promise<void> {
+    const { knex } = this;
+    const updatedAt = new Date();
+    await assignNewIds(knex, todos);
+    for await (const todo of Object.values(todos)) {
+      if (todo) {
+        const meta = todo.metadata;
+        if (todo.inserts.length > 0) {
+          await batchInsert(knex, meta, todo.inserts);
+        }
+        if (todo.updates.length > 0) {
+          todo.updates.forEach((e) => (e.__orm.data["updatedAt"] = updatedAt));
+          await batchUpdate(knex, meta, todo.updates);
+        }
+        if (todo.deletes.length > 0) {
+          await batchDelete(knex, meta, todo.deletes);
+        }
+      }
+    }
+  }
+
+  async flushJoinTables(joinRows: Record<string, JoinRowTodo>): Promise<void> {
+    const { knex } = this;
+    for await (const [joinTableName, { m2m, newRows, deletedRows }] of Object.entries(joinRows)) {
+      if (newRows.length > 0) {
+        const ids = await knex
+          .batchInsert(
+            joinTableName,
+            newRows.map((row) => {
+              // The rows in EntityManager.joinRows point to entities, change those to ints
+              const { id, created_at, m2m, ...fkColumns } = row;
+              Object.keys(fkColumns).forEach((key) => {
+                const meta = key == m2m.columnName ? getMetadata(m2m.entity) : m2m.otherMeta;
+                fkColumns[key] = keyToNumber(meta, maybeResolveReferenceToId(fkColumns[key]));
+              });
+              return fkColumns;
+            }),
+          )
+          .returning("id");
+        for (let i = 0; i < ids.length; i++) {
+          newRows[i].id = ids[i];
+        }
+      }
+      if (deletedRows.length > 0) {
+        // `remove`s that were done against unloaded ManyToManyCollections will not have row ids
+        const [haveIds, noIds] = partition(deletedRows, (r) => r.id !== -1);
+
+        if (haveIds.length > 0) {
+          await knex(joinTableName)
+            .del()
+            .whereIn(
+              "id",
+              haveIds.map((e) => e.id!),
+            );
+        }
+
+        if (noIds.length > 0) {
+          const data = noIds.map(
+            (e) =>
+              [
+                deTagIds(m2m.meta, [maybeResolveReferenceToId(e[m2m.columnName])!])[0],
+                deTagIds(m2m.otherMeta, [maybeResolveReferenceToId(e[m2m.otherColumnName])!])[0],
+              ] as any,
+          );
+          await knex(joinTableName).del().whereIn([m2m.columnName, m2m.otherColumnName], data);
+        }
       }
     }
   }
@@ -172,55 +318,7 @@ export function getTodo(todos: Record<string, Todo>, entity: Entity): Todo {
   return todo;
 }
 
-export async function flushJoinTables(knex: Knex, joinRows: Record<string, JoinRowTodo>): Promise<void> {
-  for await (const [joinTableName, { m2m, newRows, deletedRows }] of Object.entries(joinRows)) {
-    if (newRows.length > 0) {
-      const ids = await knex
-        .batchInsert(
-          joinTableName,
-          newRows.map((row) => {
-            // The rows in EntityManager.joinRows point to entities, change those to ints
-            const { id, created_at, m2m, ...fkColumns } = row;
-            Object.keys(fkColumns).forEach((key) => {
-              const meta = key == m2m.columnName ? getMetadata(m2m.entity) : m2m.otherMeta;
-              fkColumns[key] = keyToNumber(meta, maybeResolveReferenceToId(fkColumns[key]));
-            });
-            return fkColumns;
-          }),
-        )
-        .returning("id");
-      for (let i = 0; i < ids.length; i++) {
-        newRows[i].id = ids[i];
-      }
-    }
-    if (deletedRows.length > 0) {
-      // `remove`s that were done against unloaded ManyToManyCollections will not have row ids
-      const [haveIds, noIds] = partition(deletedRows, (r) => r.id !== -1);
-
-      if (haveIds.length > 0) {
-        await knex(joinTableName)
-          .del()
-          .whereIn(
-            "id",
-            haveIds.map((e) => e.id!),
-          );
-      }
-
-      if (noIds.length > 0) {
-        const data = noIds.map(
-          (e) =>
-            [
-              deTagIds(m2m.meta, [maybeResolveReferenceToId(e[m2m.columnName])!])[0],
-              deTagIds(m2m.otherMeta, [maybeResolveReferenceToId(e[m2m.otherColumnName])!])[0],
-            ] as any,
-        );
-        await knex(joinTableName).del().whereIn([m2m.columnName, m2m.otherColumnName], data);
-      }
-    }
-  }
-}
-
-interface JoinRowTodo {
+export interface JoinRowTodo {
   // Store the m2m reference (either side of the m2m, it doesn't matter which) to help tag/untag the foreign keys
   m2m: ManyToManyCollection<any, any>;
   newRows: JoinRow[];
